worked on String manipulation
Need some sort of end character. Currently max string length is as long as the lcd screen allows, and the user just taps confirm and then the program trims spaces. It would be nice to allow the user to just navigate to an end character and skip those extra button presses